activitypub: a bit of fettling of the inbox controller
